requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
langchain-community
langchain-groq
langchain-pinecone
langchain-huggingface
pinecone-client==4.1.0
sentence-transformers==3.0.1
python-dotenv==1.1.1
pydantic==2.11.7
httpx==0.28.1